api/requirements.txt
flask==2.0.1
python-dotenv==0.18.0
dnspython==2.1.0
wheel==0.36.2
flask-mongoengine==1.0.0